# swagger.yaml
swagger: "2.0"
info:
  title: Olibo League API
  description: |
    ## API de gestion de la Ligue Olibo

    Cette API permet de gérer intégralement une compétition de football : équipes, matchs,
    licences, votes, paiements, médias et bien plus.

    ### Authentification
    La plupart des endpoints nécessitent un **JWT Bearer Token**.
    Connectez-vous via `POST /api/auth/login` pour obtenir vos tokens,
    puis cliquez sur **Authorize** (🔒) et entrez : `Bearer <votre_access_token>`

    ### Rôles disponibles
    | Rôle | Description |
    |------|-------------|
    | `super_admin` | Accès total |
    | `admin_competition` | Gestion des compétitions et équipes |
    | `operator` | Gestion des matchs |
    | `referee` | Arbitre — feuilles de match et incidents |
    | `commissioner` | Commissaire de match |
    | `team_captain` | Capitaine d'équipe |
    | `coach` | Entraîneur |
    | `player` | Joueur |
    | `spectator` | Spectateur |
    | `journalist` | Journaliste |

  version: "1.0.0"
  contact:
    name: Équipe Olibo
    email: contact@olibo.com
  license:
    name: Propriétaire
host: "127.0.0.1:8000"
basePath: "/"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Format requis : **Bearer {access_token}**"
...
definitions:
...
  Match:
    type: object
    properties:
      id:
        type: integer
        example: 1
      competition_id:
        type: integer
        example: 1
      home_team_id:
        type: integer
        example: 1
      away_team_id:
        type: integer
        example: 2
      scheduled_date:
        type: string
        format: date-time
        example: "2026-03-15T15:00:00Z"
      location:
        type: string
        example: Stade Omnisports de Yaoundé
      matchday:
        type: integer
        example: 5
      status:
        type: string
        enum: [scheduled, in_progress, completed, cancelled]
        example: scheduled
      home_team_goals:
        type: integer
        example: 2
      away_team_goals:
        type: integer
        example: 1
      referee_id:
        type: integer
        example: 8
  MatchSheet:
    type: object
    properties:
      id:
        type: integer
        example: 1
      match_id:
        type: integer
        example: 1
      home_team_goals:
        type: integer
        example: 2
      away_team_goals:
        type: integer
        example: 1
      notes:
        type: string
        example: Match sans incident majeur
      is_validated:
        type: boolean
        example: false
  MatchEvent:
    type: object
    properties:
      id:
        type: integer
        example: 1
      match_id:
        type: integer
        example: 1
      member_id:
        type: integer
        example: 15
      event_type:
        type: string
        enum: [goal, assist, yellow_card, red_card, substitution]
        example: goal
      minute:
        type: integer
        minimum: 0
        maximum: 130
        example: 67
      card_type:
        type: string
        enum: [yellow, red]
        example: yellow
      notes:
        type: string
        example: Tir du droit depuis la surface
  MatchSheetFull:
    type: object
    properties:
      match:
        $ref: '#/definitions/Match'
      sheet:
        $ref: '#/definitions/MatchSheet'
      events:
        type: array
        items:
          $ref: '#/definitions/MatchEvent'

  # --- Ranking ---
  Ranking:
    type: object
    properties:
      id:
        type: integer
        example: 1
      competition_id:
        type: integer
        example: 1
      team_id:
        type: integer
        example: 1
      position:
        type: integer
        example: 1
      played:
        type: integer
        example: 10
      won:
        type: integer
        example: 7
      drawn:
        type: integer
        example: 2
      lost:
        type: integer
        example: 1
      goals_for:
        type: integer
        example: 22
      goals_against:
        type: integer
        example: 8
      goal_difference:
        type: integer
        example: 14
      points:
        type: integer
        example: 23

  # --- Voting ---
...
  Error:
    type: object
    properties:
      error:
        type: string
        example: Unauthorized
      message:
        type: string
        example: Token JWT invalide ou expiré

# ============================================================
# PATHS
# ============================================================
paths:
...
  /api/match_sheet/matches:
    post:
      tags: [Match]
      summary: Créer un match
      description: Planifie un nouveau match dans une compétition.
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [competition_id, home_team_id, away_team_id, scheduled_date]
            properties:
              competition_id:
                type: integer
                example: 1
              home_team_id:
                type: integer
                example: 1
              away_team_id:
                type: integer
                example: 2
              scheduled_date:
                type: string
                format: date-time
                example: "2026-03-15T15:00:00Z"
              matchday:
                type: integer
                example: 5
              location:
                type: string
                example: Stade Omnisports de Yaoundé
              referee_id:
                type: integer
                example: 8
      responses:
        201:
          description: Match créé
          schema:
            $ref: '#/definitions/Match'
        400:
          description: Données invalides (même équipe, competition invalide)
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Permissions insuffisantes
    get:
      tags: [Match]
      summary: Lister les matchs
      description: Accessible publiquement. Filtrable par compétition et statut.
      parameters:
        - in: query
          name: competition_id
          type: integer
          description: Filtrer par compétition
        - in: query
          name: status
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
      responses:
        200:
          description: Liste des matchs
          schema:
            type: array
            items:
              $ref: '#/definitions/Match'
  /api/match_sheet/matches/{id}:
    get:
      tags: [Match]
      summary: Obtenir un match
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Données du match
          schema:
            $ref: '#/definitions/Match'
        404:
          description: Match non trouvé
    put:
      tags: [Match]
      summary: Modifier un match
      description: Impossible si le match a le statut `completed`.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          schema:
            type: object
            properties:
              scheduled_date:
                type: string
                format: date-time
              location:
                type: string
              matchday:
                type: integer
              referee_id:
                type: integer
      responses:
        200:
          description: Match mis à jour
          schema:
            $ref: '#/definitions/Match'
        400:
          description: Match déjà complété
        403:
          description: Permissions insuffisantes
  /api/match_sheet/matches/{id}/sheet:
    get:
      tags: [Match]
      summary: Obtenir la feuille de match complète
      description: Retourne le match, la feuille et tous les événements enregistrés.
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Feuille de match complète
          schema:
            $ref: '#/definitions/MatchSheetFull'
    post:
      tags: [Match]
      summary: Remplir la feuille de match
      description: Réservé aux rôles `referee` et `commissioner`.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          schema:
            type: object
            properties:
              home_team_goals:
                type: integer
                minimum: 0
                example: 2
              away_team_goals:
                type: integer
                minimum: 0
                example: 1
              notes:
                type: string
                example: Match joué sans incident
      responses:
        200:
          description: Feuille de match remplie/mise à jour
          schema:
            $ref: '#/definitions/MatchSheet'
        403:
          description: Réservé au referee et commissioner
  /api/match_sheet/matches/{id}/sheet/validate:
    post:
      tags: [Match]
      summary: Valider la feuille de match
      description: |
        Verrouille la feuille de match (`is_validated = true`).
        Après validation, **les événements ne peuvent plus être modifiés**.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Feuille de match validée
          schema:
            $ref: '#/definitions/MatchSheet'
        403:
          description: Permissions insuffisantes
  /api/match_sheet/matches/{id}/close:
    post:
      tags: [Match]
      summary: Clôturer un match
      description: |
        Passe le statut du match à `completed` et **recalcule automatiquement
        les classements** de la compétition.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Match clôturé — classements mis à jour
          schema:
            type: object
            properties:
              match:
                $ref: '#/definitions/Match'
              rankings:
                type: array
                items:
                  $ref: '#/definitions/Ranking'
        403:
          description: Permissions insuffisantes
  /api/match_sheet/matches/{id}/events:
    post:
      tags: [Match]
      summary: Enregistrer un événement de match
      description: |
        Types d'événements : `goal`, `assist`, `yellow_card`, `red_card`, `substitution`.
        Impossible si la feuille de match est déjà **validée**.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [member_id, event_type, minute]
            properties:
              member_id:
                type: integer
                example: 15
              event_type:
                type: string
                enum: [goal, assist, yellow_card, red_card, substitution]
                example: goal
              minute:
                type: integer
                minimum: 0
                maximum: 130
                example: 67
              card_type:
                type: string
                enum: [yellow, red]
                description: Requis si event_type est yellow_card ou red_card
              notes:
                type: string
      responses:
        201:
          description: Événement enregistré
          schema:
            $ref: '#/definitions/MatchEvent'
        400:
          description: Feuille de match déjà validée
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Réservé au referee et commissioner
    get:
      tags: [Match]
      summary: Lister les événements d'un match
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Liste des événements du match
          schema:
            type: array
            items:
              $ref: '#/definitions/MatchEvent'
  /api/match_sheet/matches/{id}/events/{event_id}:
    delete:
      tags: [Match]
      summary: Supprimer un événement
      description: Impossible si la feuille de match est déjà validée.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: path
          name: event_id
          required: true
          type: integer
      responses:
        204:
          description: Événement supprimé
        400:
          description: Feuille déjà validée — suppression impossible
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Permissions insuffisantes

  # ==================== RANKING ====================
...